% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_less_na.R
\name{obs_less_na}
\alias{obs_less_na}
\title{A Summary Statistics Function}
\usage{
obs_less_na(x = TRUE, y = FALSE)
}
\arguments{
\item{x}{Name of the input object, which defaults to TRUE.}

\item{y}{Name that you wish to assign to the data frame that will be created, 
which defaults to FALSE.}
}
\description{
This function creates a data frame that lists the length of an object, how 
many observations are missing within that object, and the number of non-NA 
observations within that object.
}
\examples{
obs_less_na(iris$Sepal.Length)
obs_less_na(iris$Sepal.Length, "Sepal length data within the iris dataset.")

x <- c(1:100, NA, NA, NA)
obs_less_na(x)
obs_less_na(x, "This is an example table.")

}
\author{
Stephanie Langeland stephanie.langeland@gmail.com
}
\keyword{SLpackage}
